requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
pyyaml>=6.0
numpy>=1.21.0
scipy>=1.7.0

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.15.0
pandas>=1.5.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
requests>=2.31.0

# Logging and monitoring
tqdm>=4.65.0

# Statistical analysis
scikit-learn>=1.3.0

# Additional dependencies
python-dotenv>=1.0.0
jsonschema>=4.17.0
click>=8.1.0
rich>=13.0.0 